fix: fix server side component result page
